renamed _credential.py -> _headers_mixin.py
