+++ sdk/search/azure-search-documents/azure/search/documents/_headers_mixin.py
